Separate the ToDo class Deadline class and Event class from Task class
The 3 subclass are originally composited with Task class.

They are separated because they have different behaviours.

Let's create cleaner code!
